services/llm-backend/requirements.txt
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
axiom-py>=0.1.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
transformers>=4.35.0
torch>=2.0.0